shited the function to getContributedRepos.ts
