+++ src/popup/pages/profile.tsx
@@ -7,7 +7,8 @@ import OpenSaucedLogo from "../../assets/opensauced-logo.svg";
 import { getUserData, getUserPRData, getUserHighlightsData } from "../../utils/fetchOpenSaucedApiData";
 import { emojify } from "node-emoji";
 import { goBack, goTo } from "react-chrome-extension-router";
-import { getRelativeDays, countUniqueFullNames } from "../../utils/dateUtils";
+import { getRelativeDays } from "../../utils/dateUtils";
+import { countUniqueFullNames } from "../../utils/getContributedRepos";
 import { getUserPRVelocity } from "../../utils/getUserPRVelocity";
 import { BiExit } from "react-icons/bi";
 import Start from "./start";
